+++ pos-supply-chain/src/main/java/com/dianba/pos/supplychain/service/WarehouseOrgManager.java
@@ -2,5 +2,5 @@ package com.dianba.pos.supplychain.service;
 
 public interface WarehouseOrgManager {
 
-    int getNearbyWarehouse(double latitude, double longitude);
+    Long getNearbyWarehouse(double latitude, double longitude);
 }
